Allow us to use keystoneclient's session
The session object is a cross-client means of standardizing the
transport layer.

HTTPClient object has diverged significantly from other clients. It is
easier to simply replace it if a session is provided. If a session is
provided then users of the library need to be aware that functions such
as authenticate() will no longer have any effect/are in error because
this is no longer managed by cinderclient.

Change-Id: I8935598bda8adb014d1055e39721d8eb35b3ef50